Add test utils crate.
