+++ nextgen/project-graph/Cargo.toml
@@ -32,5 +32,6 @@ thiserror = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
+moon_test_utils2 = { path = "../test-utils" }
 starbase_sandbox = { workspace = true }
 tokio = { workspace = true }
