+++ nextgen/test-utils/src/lib.rs
@@ -1,0 +1,3 @@
+mod project_graph;
+
+pub use project_graph::*;
